+++ ui/dashboards/src/components/Variables/VariableEditorForm/index.ts
@@ -1,0 +1,1 @@
+export * from './VariableEditorForm';
